Work around leaking Devel::REPL's internal storage of values to dump
p() uses a prototype so sees @to_dump as an array ref.  This causes it
to dump all output surrounded by an array ref, which is often confusing
even to me, and I know why it happens!

Disable the use of prototypes and loop over @to_dump to process
individual values.  While use_prototypes => 0 isn't strictly necessary
with the loop (because there will never be non-ref values in an array),
it declares the code's intent.

use_prototypes requires DDP 0.17.