Always pass return values through DDP
The previous behaviours were cribbed from the default DDS plugin, but
have proved too annoying.

"Stringify if the object can" leads to much cursing for Class::DBI
objects, for example, which default to stringifying to their numeric
ids.  I want to see the actual object!

"Use a single, non-ref value as is" doesn't make much sense since we
don't do that for a list of simple, non-ref values.

As a bonus, even simple values are now color coded by data "type" to
match the more complex output of DDP.